Remove `DiscoveryBase.h`
Redistribute code from that file into various files for RTPS and Static
discovery.
