+++ dds/DCPS/RTPS/Spdp.h
@@ -1,6 +1,4 @@
 /*
- *
- *
  * Distributed under the OpenDDS License.
  * See: http://www.opendds.org/license.html
  */
@@ -23,6 +21,8 @@
 #include <dds/DCPS/MultiTask.h>
 #include <dds/DCPS/JobQueue.h>
 #include <dds/DCPS/NetworkConfigMonitor.h>
+#include <dds/DCPS/FibonacciSequence.h>
+#include <dds/DCPS/BuiltInTopicDataReaderImpls.h>
 #include <dds/DCPS/security/framework/SecurityConfig_rch.h>
 #ifdef OPENDDS_SECURITY
 #  include <dds/DCPS/security/framework/SecurityConfig.h>
@@ -35,18 +35,22 @@
 #include <dds/DdsDcpsInfoUtilsC.h>
 #include <dds/DdsDcpsCoreTypeSupportImpl.h>
 
-#ifdef ACE_HAS_CPP11
-#  include <atomic>
-#else
+#ifndef ACE_HAS_CPP11
 #  include <ace/Atomic_Op.h>
 #endif
 #include <ace/SOCK_Dgram.h>
 #include <ace/SOCK_Dgram_Mcast.h>
 #include <ace/Thread_Mutex.h>
 
-#if !defined (ACE_LACKS_PRAGMA_ONCE)
-#pragma once
-#endif /* ACE_LACKS_PRAGMA_ONCE */
+#ifdef ACE_HAS_CPP11
+#  include <atomic>
+#endif
+
+#ifndef ACE_LACKS_PRAGMA_ONCE
+#  pragma once
+#endif
+
+/* ParticipantData_t */
 
 OPENDDS_BEGIN_VERSIONED_NAMESPACE_DECL
 
@@ -62,12 +66,157 @@ const char SEDP_AGENT_INFO_KEY[] = "SEDP";
 /// Each instance of class Spdp represents the implementation of the RTPS
 /// Simple Participant Discovery Protocol for a single local DomainParticipant.
 class OpenDDS_Rtps_Export Spdp
-  : public DCPS::LocalParticipant<Sedp>
+  : public DCPS::RcObject
 #ifdef OPENDDS_SECURITY
   , public ICE::AgentInfoListener
 #endif
 {
 public:
+  struct DiscoveredParticipant {
+
+    DiscoveredParticipant()
+    : location_ih_(DDS::HANDLE_NIL)
+    , bit_ih_(DDS::HANDLE_NIL)
+    , seq_reset_count_(0)
+#ifdef OPENDDS_SECURITY
+    , have_spdp_info_(false)
+    , have_sedp_info_(false)
+    , have_auth_req_msg_(false)
+    , have_handshake_msg_(false)
+    , handshake_resend_falloff_(TimeDuration::zero_value)
+    , auth_state_(AUTH_STATE_HANDSHAKE)
+    , handshake_state_(HANDSHAKE_STATE_BEGIN_HANDSHAKE_REQUEST)
+    , is_requester_(false)
+    , auth_req_sequence_number_(0)
+    , handshake_sequence_number_(0)
+    , identity_handle_(DDS::HANDLE_NIL)
+    , handshake_handle_(DDS::HANDLE_NIL)
+    , permissions_handle_(DDS::HANDLE_NIL)
+    , extended_builtin_endpoints_(0)
+#endif
+    {
+#ifdef OPENDDS_SECURITY
+      security_info_.participant_security_attributes = 0;
+      security_info_.plugin_participant_security_attributes = 0;
+#endif
+    }
+
+    DiscoveredParticipant(
+      const ParticipantData_t& p,
+      const SequenceNumber& seq,
+      const TimeDuration& resend_period)
+    : pdata_(p)
+    , location_ih_(DDS::HANDLE_NIL)
+    , bit_ih_(DDS::HANDLE_NIL)
+    , last_seq_(seq)
+    , seq_reset_count_(0)
+#ifdef OPENDDS_SECURITY
+    , have_spdp_info_(false)
+    , have_sedp_info_(false)
+    , have_auth_req_msg_(false)
+    , have_handshake_msg_(false)
+    , handshake_resend_falloff_(resend_period)
+    , auth_state_(AUTH_STATE_HANDSHAKE)
+    , handshake_state_(HANDSHAKE_STATE_BEGIN_HANDSHAKE_REQUEST)
+    , is_requester_(false)
+    , auth_req_sequence_number_(0)
+    , handshake_sequence_number_(0)
+    , identity_handle_(DDS::HANDLE_NIL)
+    , handshake_handle_(DDS::HANDLE_NIL)
+    , permissions_handle_(DDS::HANDLE_NIL)
+    , extended_builtin_endpoints_(0)
+#endif
+    {
+      const GUID_t guid = DCPS::make_part_guid(p.participantProxy.guidPrefix);
+      assign(location_data_.guid, guid);
+      location_data_.location = 0;
+      location_data_.change_mask = 0;
+      location_data_.local_timestamp.sec = 0;
+      location_data_.local_timestamp.nanosec = 0;
+      location_data_.ice_timestamp.sec = 0;
+      location_data_.ice_timestamp.nanosec = 0;
+      location_data_.relay_timestamp.sec = 0;
+      location_data_.relay_timestamp.nanosec = 0;
+      location_data_.local6_timestamp.sec = 0;
+      location_data_.local6_timestamp.nanosec = 0;
+      location_data_.ice6_timestamp.sec = 0;
+      location_data_.ice6_timestamp.nanosec = 0;
+      location_data_.relay6_timestamp.sec = 0;
+      location_data_.relay6_timestamp.nanosec = 0;
+
+#ifdef OPENDDS_SECURITY
+      security_info_.participant_security_attributes = 0;
+      security_info_.plugin_participant_security_attributes = 0;
+#else
+      ACE_UNUSED_ARG(resend_period);
+#endif
+    }
+
+    ParticipantData_t pdata_;
+
+    struct LocationUpdate {
+      DCPS::ParticipantLocation mask_;
+      ACE_INET_Addr from_;
+      SystemTimePoint timestamp_;
+      LocationUpdate() {}
+      LocationUpdate(DCPS::ParticipantLocation mask,
+                     const ACE_INET_Addr& from,
+                     const SystemTimePoint& timestamp)
+        : mask_(mask), from_(from), timestamp_(timestamp) {}
+    };
+    typedef OPENDDS_VECTOR(LocationUpdate) LocationUpdateList;
+    LocationUpdateList location_updates_;
+    DCPS::ParticipantLocationBuiltinTopicData location_data_;
+    DDS::InstanceHandle_t location_ih_;
+
+    ACE_INET_Addr local_address_;
+    MonotonicTimePoint discovered_at_;
+    MonotonicTimePoint lease_expiration_;
+    DDS::InstanceHandle_t bit_ih_;
+    SequenceNumber last_seq_;
+    ACE_UINT16 seq_reset_count_;
+#ifdef OPENDDS_SECURITY
+    bool have_spdp_info_;
+    ICE::AgentInfo spdp_info_;
+    bool have_sedp_info_;
+    ICE::AgentInfo sedp_info_;
+    bool have_auth_req_msg_;
+    DDS::Security::ParticipantStatelessMessage auth_req_msg_;
+    bool have_handshake_msg_;
+    DDS::Security::ParticipantStatelessMessage handshake_msg_;
+    DCPS::FibonacciSequence<TimeDuration> handshake_resend_falloff_;
+    MonotonicTimePoint stateless_msg_deadline_;
+
+    MonotonicTimePoint handshake_deadline_;
+    AuthState auth_state_;
+    HandshakeState handshake_state_;
+    bool is_requester_;
+    CORBA::LongLong auth_req_sequence_number_;
+    CORBA::LongLong handshake_sequence_number_;
+
+    DDS::Security::IdentityToken identity_token_;
+    DDS::Security::PermissionsToken permissions_token_;
+    DDS::Security::PropertyQosPolicy property_qos_;
+    DDS::Security::ParticipantSecurityInfo security_info_;
+    DDS::Security::IdentityStatusToken identity_status_token_;
+    DDS::Security::IdentityHandle identity_handle_;
+    DDS::Security::HandshakeHandle handshake_handle_;
+    DDS::Security::AuthRequestMessageToken local_auth_request_token_;
+    DDS::Security::AuthRequestMessageToken remote_auth_request_token_;
+    DDS::Security::AuthenticatedPeerCredentialToken authenticated_peer_credential_token_;
+    DDS::Security::SharedSecretHandle_var shared_secret_handle_;
+    DDS::Security::PermissionsHandle permissions_handle_;
+    DDS::Security::ParticipantCryptoTokenSeq crypto_tokens_;
+    DDS::Security::ExtendedBuiltinEndpointSet_t extended_builtin_endpoints_;
+#endif
+  };
+
+  typedef OPENDDS_MAP_CMP(GUID_t, DiscoveredParticipant,
+                          GUID_tKeyLessThan) DiscoveredParticipantMap;
+  typedef typename DiscoveredParticipantMap::iterator DiscoveredParticipantIter;
+  typedef typename DiscoveredParticipantMap::const_iterator
+    DiscoveredParticipantConstIter;
+
 
   Spdp(DDS::DomainId_t domain,
        DCPS::RepoId& guid,
@@ -170,7 +319,7 @@ public:
   DDS::DomainId_t get_domain_id() const { return domain_; }
   DDS::Security::PermissionsHandle lookup_participant_permissions(const DCPS::RepoId& id) const;
 
-  DCPS::AuthState lookup_participant_auth_state(const DCPS::RepoId& id) const;
+  AuthState lookup_participant_auth_state(const GUID_t& id) const;
 
   void process_participant_ice(const ParameterList& plist,
                                const ParticipantData_t& pdata,
@@ -222,8 +371,130 @@ public:
   void get_and_reset_relay_message_counts(DCPS::RelayMessageCounts& spdp,
                                           DCPS::RelayMessageCounts& sedp);
 
+  void ignore_domain_participant(const GUID_t& ignoreId);
+
+  bool update_domain_participant_qos(const DDS::DomainParticipantQos& qos);
+
+  DCPS::TopicStatus assert_topic(GUID_t& topicId, const char* topicName,
+    const char* dataTypeName, const DDS::TopicQos& qos,
+    bool hasDcpsKey, DCPS::TopicCallbacks* topic_callbacks);
+
+  DCPS::TopicStatus find_topic(
+    const char* topicName,
+    CORBA::String_out dataTypeName,
+    DDS::TopicQos_out qos,
+    GUID_t& topicId)
+  {
+    return endpoint_manager().find_topic(topicName, dataTypeName, qos, topicId);
+  }
+
+  DCPS::TopicStatus remove_topic(const GUID_t& topicId)
+  {
+    return endpoint_manager().remove_topic(topicId);
+  }
+
+  void ignore_topic(const GUID_t& ignoreId)
+  {
+    ACE_GUARD(ACE_Thread_Mutex, g, lock_);
+    endpoint_manager().ignore(ignoreId);
+  }
+
+  bool update_topic_qos(const GUID_t& topicId, const DDS::TopicQos& qos)
+  {
+    return endpoint_manager().update_topic_qos(topicId, qos);
+  }
+
+  GUID_t add_publication(
+    const GUID_t& topicId,
+    DCPS::DataWriterCallbacks_rch publication,
+    const DDS::DataWriterQos& qos,
+    const DCPS::TransportLocatorSeq& transInfo,
+    const DDS::PublisherQos& publisherQos,
+    const XTypes::TypeInformation& type_info)
+  {
+    return endpoint_manager().add_publication(topicId, publication, qos,
+                                              transInfo, publisherQos, type_info);
+  }
+
+  void remove_publication(const GUID_t& publicationId)
+  {
+    endpoint_manager().remove_publication(publicationId);
+  }
+
+  void ignore_publication(const GUID_t& ignoreId)
+  {
+    ACE_GUARD(ACE_Thread_Mutex, g, lock_);
+    return endpoint_manager().ignore(ignoreId);
+  }
+
+  bool update_publication_qos(
+    const GUID_t& publicationId,
+    const DDS::DataWriterQos& qos,
+    const DDS::PublisherQos& publisherQos)
+  {
+    return endpoint_manager().update_publication_qos(publicationId, qos, publisherQos);
+  }
+
+  void update_publication_locators(const GUID_t& publicationId,
+                                   const DCPS::TransportLocatorSeq& transInfo)
+  {
+    endpoint_manager().update_publication_locators(publicationId, transInfo);
+  }
+
+  GUID_t add_subscription(
+    const GUID_t& topicId,
+    DCPS::DataReaderCallbacks_rch subscription,
+    const DDS::DataReaderQos& qos,
+    const DCPS::TransportLocatorSeq& transInfo,
+    const DDS::SubscriberQos& subscriberQos,
+    const char* filterClassName,
+    const char* filterExpr,
+    const DDS::StringSeq& params,
+    const XTypes::TypeInformation& type_info)
+  {
+    return endpoint_manager().add_subscription(topicId, subscription, qos, transInfo,
+      subscriberQos, filterClassName, filterExpr, params, type_info);
+  }
+
+  void remove_subscription(const GUID_t& subscriptionId)
+  {
+    endpoint_manager().remove_subscription(subscriptionId);
+  }
+
+  void ignore_subscription(const GUID_t& ignoreId)
+  {
+    ACE_GUARD(ACE_Thread_Mutex, g, lock_);
+    return endpoint_manager().ignore(ignoreId);
+  }
+
+  bool update_subscription_qos(
+    const GUID_t& subscriptionId,
+    const DDS::DataReaderQos& qos,
+    const DDS::SubscriberQos& subscriberQos)
+  {
+    return endpoint_manager().update_subscription_qos(subscriptionId, qos, subscriberQos);
+  }
+
+  bool update_subscription_params(const GUID_t& subId, const DDS::StringSeq& params)
+  {
+    return endpoint_manager().update_subscription_params(subId, params);
+  }
+
+  void update_subscription_locators(const GUID_t& subId, const DCPS::TransportLocatorSeq& transInfo)
+  {
+    endpoint_manager().update_subscription_locators(subId, transInfo);
+  }
+
+  DDS::Subscriber_var bit_subscriber() const
+  {
+    return bit_subscriber_;
+  }
+
 protected:
   Sedp& endpoint_manager() { return *sedp_; }
+
+  void remove_discovered_participant(DiscoveredParticipantIter& iter);
+
   void remove_discovered_participant_i(DiscoveredParticipantIter& iter);
 
 #ifndef DDS_HAS_MINIMUM_BIT
@@ -234,7 +505,62 @@ protected:
 
   bool announce_domain_participant_qos();
 
+  void type_lookup_service(const XTypes::TypeLookupService_rch type_lookup_service)
+  {
+    endpoint_manager().type_lookup_service(type_lookup_service);
+  }
+
 private:
+#ifndef DDS_HAS_MINIMUM_BIT
+  DCPS::ParticipantBuiltinTopicDataDataReaderImpl* part_bit()
+  {
+    DDS::Subscriber_var bit_sub(bit_subscriber());
+    if (!bit_sub.in())
+      return 0;
+
+    DDS::DataReader_var d =
+      bit_sub->lookup_datareader(DCPS::BUILT_IN_PARTICIPANT_TOPIC);
+    return dynamic_cast<DCPS::ParticipantBuiltinTopicDataDataReaderImpl*>(d.in());
+  }
+
+  DCPS::ParticipantLocationBuiltinTopicDataDataReaderImpl* part_loc_bit()
+  {
+    DDS::Subscriber_var bit_sub(bit_subscriber());
+    if (!bit_sub.in())
+      return 0;
+
+    DDS::DataReader_var d =
+      bit_sub->lookup_datareader(DCPS::BUILT_IN_PARTICIPANT_LOCATION_TOPIC);
+    return dynamic_cast<DCPS::ParticipantLocationBuiltinTopicDataDataReaderImpl*>(d.in());
+  }
+
+  DCPS::ConnectionRecordDataReaderImpl* connection_record_bit()
+  {
+    DDS::Subscriber_var bit_sub(bit_subscriber());
+    if (!bit_sub.in())
+      return 0;
+
+    DDS::DataReader_var d =
+      bit_sub->lookup_datareader(DCPS::BUILT_IN_CONNECTION_RECORD_TOPIC);
+    return dynamic_cast<DCPS::ConnectionRecordDataReaderImpl*>(d.in());
+  }
+
+  DCPS::InternalThreadBuiltinTopicDataDataReaderImpl* internal_thread_bit()
+  {
+    DDS::Subscriber_var bit_sub(bit_subscriber());
+    if (!bit_sub.in())
+      return 0;
+
+    DDS::DataReader_var d =
+      bit_sub->lookup_datareader(DCPS::BUILT_IN_INTERNAL_THREAD_TOPIC);
+    return dynamic_cast<DCPS::InternalThreadBuiltinTopicDataDataReaderImpl*>(d.in());
+  }
+#endif /* DDS_HAS_MINIMUM_BIT */
+
+#ifdef OPENDDS_SECURITY
+  typedef OPENDDS_MAP_CMP(GUID_t, DDS::Security::AuthRequestMessageToken, GUID_tKeyLessThan)
+    PendingRemoteAuthTokenMap;
+#endif
 
   void init(DDS::DomainId_t domain,
             DCPS::RepoId& guid,
@@ -242,6 +568,10 @@ private:
             RtpsDiscovery* disco,
             XTypes::TypeLookupService_rch tls);
 
+  mutable ACE_Thread_Mutex lock_;
+  DDS::Subscriber_var bit_subscriber_;
+  DDS::DomainParticipantQos qos_;
+  DiscoveredParticipantMap participants_;
   RtpsDiscovery* disco_;
   DCPS::RcHandle<RtpsDiscoveryConfig> config_;
 
@@ -550,8 +880,8 @@ private:
   friend class ::DDS_TEST;
 };
 
-}
-}
+} // namespace RTPS
+} // namespace OpenDDS
 
 OPENDDS_END_VERSIONED_NAMESPACE_DECL
