+++ dds/DCPS/StaticDiscovery.cpp
@@ -1,6 +1,7 @@
-#include "DCPS/DdsDcps_pch.h" //Only the _pch include should start with DCPS/
+#include <DCPS/DdsDcps_pch.h> // Only the _pch include should start with DCPS/
 
 #include "StaticDiscovery.h"
+
 #include "debug.h"
 #include "ConfigUtils.h"
 #include "DomainParticipantImpl.h"
@@ -10,7 +11,9 @@
 #include "Registered_Data_Types.h"
 #include "Qos_Helper.h"
 #include "DataWriterImpl.h"
+#include "DcpsUpcalls.h"
 #include "transport/framework/TransportRegistry.h"
+#include "XTypes/TypeAssignability.h"
 
 #include <ctype.h>
 
@@ -75,16 +78,28 @@ StaticEndpointManager::StaticEndpointManager(const RepoId& participant_id,
                                              ACE_Thread_Mutex& lock,
                                              const EndpointRegistry& registry,
                                              StaticParticipant& participant)
-  : EndpointManager<StaticDiscoveredParticipantData>(participant_id, lock)
+  : lock_(lock)
+  , participant_id_(participant_id)
+  , publication_counter_(0)
+  , subscription_counter_(0)
+  , topic_counter_(0)
   , registry_(registry)
 #ifndef DDS_HAS_MINIMUM_BIT
   , participant_(participant)
 #endif
+  , max_type_lookup_service_reply_period_(0)
+  , type_lookup_service_sequence_number_(0)
+  , use_xtypes_(true)
 {
 #ifdef DDS_HAS_MINIMUM_BIT
   ACE_UNUSED_ARG(participant);
 #endif
   type_lookup_init(TheServiceParticipant->interceptor());
+}
+
+StaticEndpointManager::~StaticEndpointManager()
+{
+  type_lookup_fini();
 }
 
 void StaticEndpointManager::init_bit()
@@ -288,7 +303,7 @@ StaticEndpointManager::update_subscription_params(const RepoId& /*subId*/,
 }
 
 bool
-StaticEndpointManager::disassociate(StaticDiscoveredParticipantData& /*pdata*/)
+StaticEndpointManager::disassociate()
 {
   ACE_DEBUG((LM_NOTICE, ACE_TEXT("(%P|%t) StaticEndpointManager::disassociate TODO\n")));
   // TODO
@@ -421,9 +436,8 @@ StaticEndpointManager::add_subscription_i(const RepoId& readerid,
   return DDS::RETCODE_OK;
 }
 
-DDS::ReturnCode_t
-StaticEndpointManager::remove_subscription_i(const RepoId& readerid,
-                                             LocalSubscription& sub)
+DDS::ReturnCode_t StaticEndpointManager::remove_subscription_i(
+  const GUID_t& readerid, LocalSubscription& sub)
 {
   EndpointRegistry::ReaderMapType::const_iterator pos = registry_.reader_map.find(readerid);
   if (pos == registry_.reader_map.end()) {
@@ -585,8 +599,1083 @@ StaticEndpointManager::sub_bit()
 }
 #endif /* DDS_HAS_MINIMUM_BIT */
 
+void StaticEndpointManager::type_lookup_init(ReactorInterceptor_rch reactor_interceptor)
+{
+  if (!type_lookup_reply_deadline_processor_) {
+    type_lookup_reply_deadline_processor_ = DCPS::make_rch<StaticEndpointManagerSporadic>(
+      reactor_interceptor, ref(*this), &StaticEndpointManager::remove_expired_endpoints);
+  }
+}
+
+void StaticEndpointManager::type_lookup_fini()
+{
+  if (type_lookup_reply_deadline_processor_) {
+    type_lookup_reply_deadline_processor_->cancel_and_wait();
+    type_lookup_reply_deadline_processor_.reset();
+  }
+}
+
+void StaticEndpointManager::type_lookup_service(
+  const XTypes::TypeLookupService_rch type_lookup_service)
+{
+  type_lookup_service_ = type_lookup_service;
+}
+
+void StaticEndpointManager::purge_dead_topic(const String& topic_name)
+{
+  TopicDetailsMap::iterator top_it = topics_.find(topic_name);
+  topic_names_.erase(top_it->second.topic_id());
+  topics_.erase(top_it);
+}
+
+void StaticEndpointManager::ignore(const GUID_t& to_ignore)
+{
+  // Locked prior to call from Spdp.
+  ignored_guids_.insert(to_ignore);
+  {
+    const DiscoveredPublicationIter iter = discovered_publications_.find(to_ignore);
+    if (iter != discovered_publications_.end()) {
+      // clean up tracking info
+      const String topic_name = iter->second.get_topic_name();
+      TopicDetails& td = topics_[topic_name];
+      td.remove_discovered_publication(to_ignore);
+      remove_from_bit(iter->second);
+      discovered_publications_.erase(iter);
+      // break associations
+      match_endpoints(to_ignore, td, true /*remove*/);
+      if (td.is_dead()) {
+        purge_dead_topic(topic_name);
+      }
+      return;
+    }
+  }
+  {
+    const DiscoveredSubscriptionIter iter =
+      discovered_subscriptions_.find(to_ignore);
+    if (iter != discovered_subscriptions_.end()) {
+      // clean up tracking info
+      const String topic_name = iter->second.get_topic_name();
+      TopicDetails& td = topics_[topic_name];
+      td.remove_discovered_publication(to_ignore);
+      remove_from_bit(iter->second);
+      discovered_subscriptions_.erase(iter);
+      // break associations
+      match_endpoints(to_ignore, td, true /*remove*/);
+      if (td.is_dead()) {
+        purge_dead_topic(topic_name);
+      }
+      return;
+    }
+  }
+  {
+    const OPENDDS_MAP_CMP(GUID_t, OPENDDS_STRING, GUID_tKeyLessThan)::iterator
+      iter = topic_names_.find(to_ignore);
+    if (iter != topic_names_.end()) {
+      ignored_topics_.insert(iter->second);
+      // Remove all publications and subscriptions on this topic
+      TopicDetails& td = topics_[iter->second];
+      {
+        const RepoIdSet ids = td.discovered_publications();
+        for (RepoIdSet::const_iterator ep = ids.begin(); ep!= ids.end(); ++ep) {
+          match_endpoints(*ep, td, true /*remove*/);
+          td.remove_discovered_publication(*ep);
+          // TODO: Do we need to remove from discovered_subscriptions?
+          if (shutting_down()) { return; }
+        }
+      }
+      {
+        const RepoIdSet ids = td.discovered_subscriptions();
+        for (RepoIdSet::const_iterator ep = ids.begin(); ep!= ids.end(); ++ep) {
+          match_endpoints(*ep, td, true /*remove*/);
+          td.remove_discovered_subscription(*ep);
+          // TODO: Do we need to remove from discovered_publications?
+          if (shutting_down()) { return; }
+        }
+      }
+      if (td.is_dead()) {
+        purge_dead_topic(iter->second);
+      }
+    }
+  }
+}
+
+bool StaticEndpointManager::ignoring(const GUID_t& guid) const
+{
+  return ignored_guids_.count(guid);
+}
+bool StaticEndpointManager::ignoring(const char* topic_name) const
+{
+  return ignored_topics_.count(topic_name);
+}
+
+TopicStatus StaticEndpointManager::assert_topic(
+  GUID_t& topicId, const char* topicName,
+  const char* dataTypeName, const DDS::TopicQos& qos,
+  bool hasDcpsKey, TopicCallbacks* topic_callbacks)
+{
+  ACE_GUARD_RETURN(ACE_Thread_Mutex, g, lock_, INTERNAL_ERROR);
+  TopicDetailsMap::iterator iter = topics_.find(topicName);
+  if (iter != topics_.end()) {
+    if (iter->second.local_is_set() && iter->second.local_data_type_name() != dataTypeName) {
+      return CONFLICTING_TYPENAME;
+    }
+    topicId = iter->second.topic_id();
+    iter->second.set_local(dataTypeName, qos, hasDcpsKey, topic_callbacks);
+    return FOUND;
+  }
+
+  TopicDetails& td = topics_[topicName];
+  topicId = make_topic_guid();
+  td.init(topicName, topicId);
+  topic_names_[topicId] = topicName;
+  td.set_local(dataTypeName, qos, hasDcpsKey, topic_callbacks);
+
+  return CREATED;
+}
+
+TopicStatus StaticEndpointManager::find_topic(
+  const char* topicName,
+  CORBA::String_out dataTypeName,
+  DDS::TopicQos_out qos,
+  GUID_t& topicId)
+{
+  ACE_GUARD_RETURN(ACE_Thread_Mutex, g, lock_, INTERNAL_ERROR);
+  TopicDetailsMap::const_iterator iter = topics_.find(topicName);
+  if (iter == topics_.end()) {
+    return NOT_FOUND;
+  }
+
+  const TopicDetails& td = iter->second;
+
+  dataTypeName = td.local_data_type_name().c_str();
+  qos = new DDS::TopicQos(td.local_qos());
+  topicId = td.topic_id();
+  return FOUND;
+}
+
+TopicStatus StaticEndpointManager::remove_topic(const GUID_t& topicId)
+{
+  ACE_GUARD_RETURN(ACE_Thread_Mutex, g, lock_, INTERNAL_ERROR);
+  TopicNameMap::iterator name_iter = topic_names_.find(topicId);
+  if (name_iter == topic_names_.end()) {
+    return NOT_FOUND;
+  }
+  const String& name = name_iter->second;
+  TopicDetails& td = topics_[name];
+  td.unset_local();
+  if (td.is_dead()) {
+    purge_dead_topic(name);
+  }
+
+  return REMOVED;
+}
+
+GUID_t StaticEndpointManager::add_publication(
+  const GUID_t& topicId,
+  DataWriterCallbacks_rch publication,
+  const DDS::DataWriterQos& qos,
+  const TransportLocatorSeq& transInfo,
+  const DDS::PublisherQos& publisherQos,
+  const XTypes::TypeInformation& type_info)
+{
+  ACE_GUARD_RETURN(ACE_Thread_Mutex, g, lock_, GUID_t());
+
+  GUID_t rid = participant_id_;
+  assign_publication_key(rid, topicId, qos);
+  LocalPublication& pb = local_publications_[rid];
+  pb.topic_id_ = topicId;
+  pb.publication_ = publication;
+  pb.qos_ = qos;
+  pb.trans_info_ = transInfo;
+  pb.publisher_qos_ = publisherQos;
+  pb.type_info_ = type_info;
+  const OPENDDS_STRING& topic_name = topic_names_[topicId];
+
+  TopicDetails& td = topics_[topic_name];
+  td.add_local_publication(rid);
+
+  if (DDS::RETCODE_OK != add_publication_i(rid, pb)) {
+    return GUID_t();
+  }
+
+  if (DDS::RETCODE_OK != write_publication_data(rid, pb)) {
+    return GUID_t();
+  }
+
+  if (DCPS_debug_level > 3) {
+    ACE_DEBUG((LM_DEBUG, ACE_TEXT("(%P|%t) StaticEndpointManager::add_publication - ")
+               ACE_TEXT("calling match_endpoints\n")));
+  }
+  match_endpoints(rid, td);
+
+  return rid;
+}
+
+void StaticEndpointManager::remove_publication(const GUID_t& publicationId)
+{
+  ACE_GUARD(ACE_Thread_Mutex, g, lock_);
+  LocalPublicationIter iter = local_publications_.find(publicationId);
+  if (iter != local_publications_.end()) {
+    if (DDS::RETCODE_OK == remove_publication_i(publicationId, iter->second)) {
+      OPENDDS_STRING topic_name = topic_names_[iter->second.topic_id_];
+      local_publications_.erase(publicationId);
+      TopicDetailsMap::iterator top_it = topics_.find(topic_name);
+      if (top_it != topics_.end()) {
+        match_endpoints(publicationId, top_it->second, true /*remove*/);
+        top_it->second.remove_local_publication(publicationId);
+        // Local, no need to check for dead topic.
+      }
+    } else {
+      ACE_ERROR((LM_ERROR,
+                 ACE_TEXT("(%P|%t) ERROR: StaticEndpointManager::remove_publication - ")
+                 ACE_TEXT("Failed to publish dispose msg\n")));
+    }
+  }
+}
+
+void StaticEndpointManager::update_publication_locators(
+  const GUID_t& publicationId, const TransportLocatorSeq& transInfo)
+{
+  ACE_GUARD(ACE_Thread_Mutex, g, lock_);
+  LocalPublicationIter iter = local_publications_.find(publicationId);
+  if (iter != local_publications_.end()) {
+    if (DCPS_debug_level > 3) {
+      const GuidConverter conv(publicationId);
+      ACE_DEBUG((LM_INFO,
+        ACE_TEXT("(%P|%t) StaticEndpointManager::update_publication_locators - updating locators for %C\n"),
+        OPENDDS_STRING(conv).c_str()));
+    }
+    iter->second.trans_info_ = transInfo;
+    write_publication_data(publicationId, iter->second);
+  }
+}
+
+GUID_t StaticEndpointManager::add_subscription(
+  const GUID_t& topicId,
+  DataReaderCallbacks_rch subscription,
+  const DDS::DataReaderQos& qos,
+  const TransportLocatorSeq& transInfo,
+  const DDS::SubscriberQos& subscriberQos,
+  const char* filterClassName,
+  const char* filterExpr,
+  const DDS::StringSeq& params,
+  const XTypes::TypeInformation& type_info)
+{
+  ACE_GUARD_RETURN(ACE_Thread_Mutex, g, lock_, GUID_t());
+
+  GUID_t rid = participant_id_;
+  assign_subscription_key(rid, topicId, qos);
+  LocalSubscription& sb = local_subscriptions_[rid];
+  sb.topic_id_ = topicId;
+  sb.subscription_ = subscription;
+  sb.qos_ = qos;
+  sb.trans_info_ = transInfo;
+  sb.subscriber_qos_ = subscriberQos;
+  sb.filterProperties.filterClassName = filterClassName;
+  sb.filterProperties.filterExpression = filterExpr;
+  sb.filterProperties.expressionParameters = params;
+  sb.type_info_ = type_info;
+  const OPENDDS_STRING& topic_name = topic_names_[topicId];
+
+  TopicDetails& td = topics_[topic_name];
+  td.add_local_subscription(rid);
+
+  if (DDS::RETCODE_OK != add_subscription_i(rid, sb)) {
+    return GUID_t();
+  }
+
+  if (DDS::RETCODE_OK != write_subscription_data(rid, sb)) {
+    return GUID_t();
+  }
+
+  if (DCPS_debug_level > 3) {
+    ACE_DEBUG((LM_DEBUG, ACE_TEXT("(%P|%t) StaticEndpointManager::add_subscription - ")
+               ACE_TEXT("calling match_endpoints\n")));
+  }
+  match_endpoints(rid, td);
+
+  return rid;
+}
+
+void StaticEndpointManager::remove_subscription(const GUID_t& subscriptionId)
+{
+  ACE_GUARD(ACE_Thread_Mutex, g, lock_);
+  LocalSubscriptionIter iter = local_subscriptions_.find(subscriptionId);
+  if (iter != local_subscriptions_.end()) {
+    if (DDS::RETCODE_OK == remove_subscription_i(subscriptionId, iter->second)) {
+      OPENDDS_STRING topic_name = topic_names_[iter->second.topic_id_];
+      local_subscriptions_.erase(subscriptionId);
+      TopicDetailsMap::iterator top_it = topics_.find(topic_name);
+      if (top_it != topics_.end()) {
+        match_endpoints(subscriptionId, top_it->second, true /*remove*/);
+        top_it->second.remove_local_subscription(subscriptionId);
+        // Local, no need to check for dead topic.
+      }
+    } else {
+      ACE_ERROR((LM_ERROR,
+                 ACE_TEXT("(%P|%t) ERROR: StaticEndpointManager::remove_subscription - ")
+                 ACE_TEXT("Failed to publish dispose msg\n")));
+    }
+  }
+}
+
+void StaticEndpointManager::update_subscription_locators(
+  const GUID_t& subscriptionId,
+  const TransportLocatorSeq& transInfo)
+{
+  ACE_GUARD(ACE_Thread_Mutex, g, lock_);
+  LocalSubscriptionIter iter = local_subscriptions_.find(subscriptionId);
+  if (iter != local_subscriptions_.end()) {
+    if (DCPS_debug_level > 3) {
+      const GuidConverter conv(subscriptionId);
+      ACE_DEBUG((LM_INFO,
+        ACE_TEXT("(%P|%t) StaticEndpointManager::update_subscription_locators updating locators for %C\n"),
+        OPENDDS_STRING(conv).c_str()));
+    }
+    iter->second.trans_info_ = transInfo;
+    write_subscription_data(subscriptionId, iter->second);
+  }
+}
+
+// TODO: This is perhaps too generic since the context probably has the details this function computes.
+void StaticEndpointManager::match_endpoints(
+  GUID_t repoId, const TopicDetails& td, bool remove)
+{
+  if (DCPS_debug_level >= 4) {
+    ACE_DEBUG((LM_DEBUG, "(%P|%t) StaticEndpointManager::match_endpoints %C%C\n",
+      remove ? "remove " : "", LogGuid(repoId).c_str()));
+  }
+
+  const bool reader = GuidConverter(repoId).isReader();
+  // Copy the endpoint set - lock can be released in match()
+  RepoIdSet local_endpoints;
+  RepoIdSet discovered_endpoints;
+  if (reader) {
+    local_endpoints = td.local_publications();
+    discovered_endpoints = td.discovered_publications();
+  } else {
+    local_endpoints = td.local_subscriptions();
+    discovered_endpoints = td.discovered_subscriptions();
+  }
+
+  const bool is_remote = !equal_guid_prefixes(repoId, participant_id_);
+  if (is_remote && local_endpoints.empty()) {
+    // Nothing to match.
+    return;
+  }
+
+  for (RepoIdSet::const_iterator iter = local_endpoints.begin();
+       iter != local_endpoints.end(); ++iter) {
+    // check to make sure it's a Reader/Writer or Writer/Reader match
+    if (GuidConverter(*iter).isReader() != reader) {
+      if (remove) {
+        remove_assoc(*iter, repoId);
+      } else {
+        match(reader ? *iter : repoId, reader ? repoId : *iter);
+      }
+    }
+  }
+
+  // Remote/remote matches are a waste of time
+  if (is_remote) {
+    return;
+  }
+
+  for (RepoIdSet::const_iterator iter = discovered_endpoints.begin();
+       iter != discovered_endpoints.end(); ++iter) {
+    // check to make sure it's a Reader/Writer or Writer/Reader match
+    if (GuidConverter(*iter).isReader() != reader) {
+      if (remove) {
+        remove_assoc(*iter, repoId);
+      } else {
+        match(reader ? *iter : repoId, reader ? repoId : *iter);
+      }
+    }
+  }
+}
+
+void StaticEndpointManager::remove_assoc(const GUID_t& remove_from, const GUID_t& removing)
+{
+  if (GuidConverter(remove_from).isReader()) {
+    const LocalSubscriptionIter lsi = local_subscriptions_.find(remove_from);
+    if (lsi != local_subscriptions_.end()) {
+      lsi->second.matched_endpoints_.erase(removing);
+      const DiscoveredPublicationIter dpi = discovered_publications_.find(removing);
+      if (dpi != discovered_publications_.end()) {
+        dpi->second.matched_endpoints_.erase(remove_from);
+      }
+      WriterIdSeq writer_seq(1);
+      writer_seq.length(1);
+      writer_seq[0] = removing;
+      const size_t count = lsi->second.remote_expectant_opendds_associations_.erase(removing);
+      DataReaderCallbacks_rch drr = lsi->second.subscription_.lock();
+      if (drr) {
+        drr->remove_associations(writer_seq, false /*notify_lost*/);
+      }
+      remove_assoc_i(remove_from, lsi->second, removing);
+      // Update writer
+      if (count) {
+        write_subscription_data(remove_from, lsi->second);
+      }
+    }
+
+  } else {
+    const LocalPublicationIter lpi = local_publications_.find(remove_from);
+    if (lpi != local_publications_.end()) {
+      lpi->second.matched_endpoints_.erase(removing);
+      const DiscoveredSubscriptionIter dsi = discovered_subscriptions_.find(removing);
+      if (dsi != discovered_subscriptions_.end()) {
+        dsi->second.matched_endpoints_.erase(remove_from);
+      }
+      ReaderIdSeq reader_seq(1);
+      reader_seq.length(1);
+      reader_seq[0] = removing;
+      lpi->second.remote_expectant_opendds_associations_.erase(removing);
+      DataWriterCallbacks_rch dwr = lpi->second.publication_.lock();
+      if (dwr) {
+        dwr->remove_associations(reader_seq, false /*notify_lost*/);
+      }
+      remove_assoc_i(remove_from, lpi->second, removing);
+    }
+  }
+}
+
+void StaticEndpointManager::match(const GUID_t& writer, const GUID_t& reader)
+{
+  if (DCPS_debug_level >= 4) {
+    ACE_DEBUG((LM_DEBUG, "(%P|%t) StaticEndpointManager::match: w: %C r: %C\n",
+      LogGuid(writer).c_str(), LogGuid(reader).c_str()));
+  }
+
+  // 1. collect type info about the writer, which may be local or discovered
+  XTypes::TypeInformation* writer_type_info = 0;
+
+  const LocalPublicationIter lpi = local_publications_.find(writer);
+  DiscoveredPublicationIter dpi;
+  bool writer_local = false;
+  if (lpi != local_publications_.end()) {
+    writer_local = true;
+    writer_type_info = &lpi->second.type_info_;
+  } else if ((dpi = discovered_publications_.find(writer))
+             != discovered_publications_.end()) {
+    writer_type_info = &dpi->second.type_info_;
+  } else {
+    if (DCPS_debug_level >= 4) {
+      ACE_DEBUG((LM_DEBUG, "(%P|%t) StaticEndpointManager::match: Undiscovered Writer\n"));
+    }
+    return; // Possible and ok, since lock is released
+  }
+
+  // 2. collect type info about the reader, which may be local or discovered
+  XTypes::TypeInformation* reader_type_info = 0;
+
+  const LocalSubscriptionIter lsi = local_subscriptions_.find(reader);
+  DiscoveredSubscriptionIter dsi;
+  bool reader_local = false;
+  if (lsi != local_subscriptions_.end()) {
+    reader_local = true;
+    reader_type_info = &lsi->second.type_info_;
+  } else if ((dsi = discovered_subscriptions_.find(reader))
+             != discovered_subscriptions_.end()) {
+    reader_type_info = &dsi->second.type_info_;
+  } else {
+    if (DCPS_debug_level >= 4) {
+      ACE_DEBUG((LM_DEBUG, "(%P|%t) StaticEndpointManager::match: Undiscovered Reader\n"));
+    }
+    return; // Possible and ok, since lock is released
+  }
+
+  MatchingData md;
+
+  // if the type object is not in cache, send RPC request
+  md.writer = writer;
+  md.reader = reader;
+  md.time_added_to_map = MonotonicTimePoint::now();
+
+  if ((writer_type_info->minimal.typeid_with_size.type_id.kind() != XTypes::TK_NONE) &&
+      (reader_type_info->minimal.typeid_with_size.type_id.kind() != XTypes::TK_NONE)) {
+    if (!writer_local && reader_local) {
+      if (type_lookup_service_ &&
+          !type_lookup_service_->type_object_in_cache(
+            writer_type_info->minimal.typeid_with_size.type_id)) {
+        if (DCPS_debug_level >= 4) {
+          ACE_DEBUG((LM_DEBUG, "(%P|%t) StaticEndpointManager::match: Remote Writer\n"));
+        }
+        bool is_discovery_protected = false;
+        save_matching_data_and_get_typeobjects(
+          writer_type_info, md, MatchingPair(writer, reader), writer, is_discovery_protected);
+        return;
+      }
+    } else if (!reader_local && writer_local) {
+      if (type_lookup_service_ && !type_lookup_service_->type_object_in_cache(
+          reader_type_info->minimal.typeid_with_size.type_id)) {
+        if (DCPS_debug_level >= 4) {
+          ACE_DEBUG((LM_DEBUG, "(%P|%t) StaticEndpointManager::match: Remote Reader\n"));
+        }
+        bool is_discovery_protected = false;
+        save_matching_data_and_get_typeobjects(
+          reader_type_info, md, MatchingPair(writer, reader), reader, is_discovery_protected);
+        return;
+      }
+    }
+  }
+
+  match_continue(writer, reader);
+}
+
+void StaticEndpointManager::remove_expired_endpoints(
+  const MonotonicTimePoint& /*now*/)
+{
+  ACE_GUARD(ACE_Thread_Mutex, g, lock_);
+  const MonotonicTimePoint now = MonotonicTimePoint::now();
+
+  MatchingDataIter end_iter = matching_data_buffer_.end();
+  for (MatchingDataIter iter = matching_data_buffer_.begin(); iter != end_iter; ) {
+    // Do not try to simplify increment: "associative container erase idiom"
+    if (now - iter->second.time_added_to_map >= max_type_lookup_service_reply_period_) {
+      if (DCPS_debug_level >= 4) {
+        ACE_DEBUG((LM_DEBUG, "(%P|%t) StaticEndpointManager::remove_expired_endpoints: "
+          "clean up pending pair w: %C r: %C\n",
+          LogGuid(iter->second.writer).c_str(), LogGuid(iter->second.reader).c_str()));
+      }
+      matching_data_buffer_.erase(iter++);
+    } else {
+      ++iter;
+    }
+  }
+
+  // Clean up internal data used by getTypeDependencies
+  for (typename OrigSeqNumberMap::iterator it = orig_seq_numbers_.begin(); it != orig_seq_numbers_.end();) {
+    if (now - it->second.time_started >= max_type_lookup_service_reply_period_) {
+      if (DCPS_debug_level >= 4) {
+        ACE_DEBUG((LM_DEBUG, "(%P|%t) StaticEndpointManager::remove_expired_endpoints: "
+          "clean up type lookup data for %C\n",
+          LogGuid(it->second.participant).c_str()));
+      }
+      cleanup_type_lookup_data(it->second.participant, it->second.type_id, it->second.secure);
+      orig_seq_numbers_.erase(it++);
+    } else {
+      ++it;
+    }
+  }
+}
+
+/// This assumes that lock_ is being held
+void StaticEndpointManager::match_continue(SequenceNumber rpc_sequence_number)
+{
+  if (DCPS_debug_level >= 4) {
+    ACE_DEBUG((LM_DEBUG, "(%P|%t) StaticEndpointManager::match_continue: rpc seq: %q\n",
+      rpc_sequence_number.getValue()));
+  }
+
+  MatchingDataIter it;
+  for (it = matching_data_buffer_.begin(); it != matching_data_buffer_.end(); ++it) {
+    if (DCPS_debug_level >= 4) {
+      ACE_DEBUG((LM_DEBUG, "(%P|%t) StaticEndpointManager::match_continue: matches %q?\n",
+        it->second.rpc_sequence_number.getValue()));
+    }
+    if (it->second.rpc_sequence_number == rpc_sequence_number) {
+      const GUID_t& reader = it->second.reader;
+      const GUID_t& writer = it->second.writer;
+      matching_data_buffer_.erase(MatchingPair(writer, reader));
+      return match_continue(writer, reader);
+    }
+  }
+  if (DCPS_debug_level) {
+    ACE_ERROR((LM_ERROR, "(%P|%t) StaticEndpointManager::match_continue: "
+      " rpc seq: %q: No data found in matching data buffer\n",
+      rpc_sequence_number.getValue()));
+  }
+}
+
+void StaticEndpointManager::match_continue(const GUID_t& writer, const GUID_t& reader)
+{
+  if (DCPS_debug_level >= 4) {
+    ACE_DEBUG((LM_DEBUG, "(%P|%t) StaticEndpointManager::match_continue: w: %C r: %C\n",
+      LogGuid(writer).c_str(), LogGuid(reader).c_str()));
+  }
+
+  // 0. For discovered endpoints, we'll have the QoS info in the form of the
+  // publication or subscription BIT data which doesn't use the same structures
+  // for QoS.  In those cases we can copy the individual QoS policies to temp
+  // QoS structs:
+  DDS::DataWriterQos tempDwQos;
+  DDS::PublisherQos tempPubQos;
+  DDS::DataReaderQos tempDrQos;
+  DDS::SubscriberQos tempSubQos;
+  ContentFilterProperty_t tempCfp;
+
+  DiscoveredPublicationIter dpi = discovered_publications_.find(writer);
+  DiscoveredSubscriptionIter dsi = discovered_subscriptions_.find(reader);
+  if (dpi != discovered_publications_.end() && dsi != discovered_subscriptions_.end()) {
+    // This is a discovered/discovered match, nothing for us to do
+    return;
+  }
+
+  // 1. Collect details about the writer, which may be local or discovered
+  const DDS::DataWriterQos* dwQos = 0;
+  const DDS::PublisherQos* pubQos = 0;
+  TransportLocatorSeq* wTls = 0;
+  ACE_CDR::ULong wTransportContext = 0;
+  XTypes::TypeInformation* writer_type_info = 0;
+  OPENDDS_STRING topic_name;
+  MonotonicTime_t writer_participant_discovered_at;
+
+  const LocalPublicationIter lpi = local_publications_.find(writer);
+  bool writer_local = false, already_matched = false;
+  if (lpi != local_publications_.end()) {
+    writer_local = true;
+    dwQos = &lpi->second.qos_;
+    pubQos = &lpi->second.publisher_qos_;
+    wTls = &lpi->second.trans_info_;
+    wTransportContext = lpi->second.transport_context_;
+    already_matched = lpi->second.matched_endpoints_.count(reader);
+    writer_type_info = &lpi->second.type_info_;
+    topic_name = topic_names_[lpi->second.topic_id_];
+    writer_participant_discovered_at = lpi->second.participant_discovered_at_;
+  } else if (dpi != discovered_publications_.end()) {
+    wTls = &dpi->second.writer_data_.writerProxy.allLocators;
+    wTransportContext = dpi->second.transport_context_;
+    writer_type_info = &dpi->second.type_info_;
+    topic_name = dpi->second.get_topic_name();
+    writer_participant_discovered_at = dpi->second.participant_discovered_at_;
+
+    const DDS::PublicationBuiltinTopicData& bit =
+      dpi->second.writer_data_.ddsPublicationData;
+    tempDwQos.durability = bit.durability;
+    tempDwQos.durability_service = bit.durability_service;
+    tempDwQos.deadline = bit.deadline;
+    tempDwQos.latency_budget = bit.latency_budget;
+    tempDwQos.liveliness = bit.liveliness;
+    tempDwQos.reliability = bit.reliability;
+    tempDwQos.destination_order = bit.destination_order;
+    tempDwQos.history = TheServiceParticipant->initial_HistoryQosPolicy();
+    tempDwQos.resource_limits =
+      TheServiceParticipant->initial_ResourceLimitsQosPolicy();
+    tempDwQos.transport_priority =
+      TheServiceParticipant->initial_TransportPriorityQosPolicy();
+    tempDwQos.lifespan = bit.lifespan;
+    tempDwQos.user_data = bit.user_data;
+    tempDwQos.ownership = bit.ownership;
+    tempDwQos.ownership_strength = bit.ownership_strength;
+    tempDwQos.writer_data_lifecycle =
+      TheServiceParticipant->initial_WriterDataLifecycleQosPolicy();
+    tempDwQos.representation = bit.representation;
+    dwQos = &tempDwQos;
+
+    tempPubQos.presentation = bit.presentation;
+    tempPubQos.partition = bit.partition;
+    tempPubQos.group_data = bit.group_data;
+    tempPubQos.entity_factory =
+      TheServiceParticipant->initial_EntityFactoryQosPolicy();
+    pubQos = &tempPubQos;
+
+    populate_transport_locator_sequence(wTls, dpi, writer);
+  } else {
+    return; // Possible and ok, since lock is released
+  }
+
+  // 2. Collect details about the reader, which may be local or discovered
+  const DDS::DataReaderQos* drQos = 0;
+  const DDS::SubscriberQos* subQos = 0;
+  TransportLocatorSeq* rTls = 0;
+  ACE_CDR::ULong rTransportContext = 0;
+  const ContentFilterProperty_t* cfProp = 0;
+  XTypes::TypeInformation* reader_type_info = 0;
+  MonotonicTime_t reader_participant_discovered_at;
+
+  const LocalSubscriptionIter lsi = local_subscriptions_.find(reader);
+  bool reader_local = false;
+  if (lsi != local_subscriptions_.end()) {
+    reader_local = true;
+    drQos = &lsi->second.qos_;
+    subQos = &lsi->second.subscriber_qos_;
+    rTls = &lsi->second.trans_info_;
+    rTransportContext = lsi->second.transport_context_;
+    reader_type_info = &lsi->second.type_info_;
+    if (lsi->second.filterProperties.filterExpression[0] != 0) {
+      tempCfp.filterExpression = lsi->second.filterProperties.filterExpression;
+      tempCfp.expressionParameters = lsi->second.filterProperties.expressionParameters;
+    }
+    cfProp = &tempCfp;
+    if (!already_matched) {
+      already_matched = lsi->second.matched_endpoints_.count(writer);
+    }
+    reader_participant_discovered_at = lsi->second.participant_discovered_at_;
+  } else if (dsi != discovered_subscriptions_.end()) {
+    rTls = &dsi->second.reader_data_.readerProxy.allLocators;
+
+    populate_transport_locator_sequence(rTls, dsi, reader);
+    rTransportContext = dsi->second.transport_context_;
+
+    const DDS::SubscriptionBuiltinTopicData& bit =
+      dsi->second.reader_data_.ddsSubscriptionData;
+    tempDrQos.durability = bit.durability;
+    tempDrQos.deadline = bit.deadline;
+    tempDrQos.latency_budget = bit.latency_budget;
+    tempDrQos.liveliness = bit.liveliness;
+    tempDrQos.reliability = bit.reliability;
+    tempDrQos.destination_order = bit.destination_order;
+    tempDrQos.history = TheServiceParticipant->initial_HistoryQosPolicy();
+    tempDrQos.resource_limits =
+      TheServiceParticipant->initial_ResourceLimitsQosPolicy();
+    tempDrQos.user_data = bit.user_data;
+    tempDrQos.ownership = bit.ownership;
+    tempDrQos.time_based_filter = bit.time_based_filter;
+    tempDrQos.reader_data_lifecycle =
+      TheServiceParticipant->initial_ReaderDataLifecycleQosPolicy();
+    tempDrQos.representation = bit.representation;
+    tempDrQos.type_consistency = bit.type_consistency;
+    drQos = &tempDrQos;
+
+    tempSubQos.presentation = bit.presentation;
+    tempSubQos.partition = bit.partition;
+    tempSubQos.group_data = bit.group_data;
+    tempSubQos.entity_factory =
+      TheServiceParticipant->initial_EntityFactoryQosPolicy();
+    subQos = &tempSubQos;
+
+    cfProp = &dsi->second.reader_data_.contentFilterProperty;
+    reader_type_info = &dsi->second.type_info_;
+    reader_participant_discovered_at = dsi->second.participant_discovered_at_;
+  } else {
+    return; // Possible and ok, since lock is released
+  }
+
+  // 3. Perform type consistency check (XTypes 1.3, Section 7.6.3.4.2)
+  bool consistent = false;
+
+  TopicDetailsMap::iterator td_iter = topics_.find(topic_name);
+  if (td_iter == topics_.end()) {
+    ACE_ERROR((LM_ERROR,
+              ACE_TEXT("(%P|%t) StaticEndpointManager::match_continue - ERROR ")
+              ACE_TEXT("Didn't find topic for consistency check\n")));
+    return;
+  } else {
+    const XTypes::TypeIdentifier& writer_type_id = writer_type_info->minimal.typeid_with_size.type_id;
+    const XTypes::TypeIdentifier& reader_type_id = reader_type_info->minimal.typeid_with_size.type_id;
+    if (writer_type_id.kind() != XTypes::TK_NONE && reader_type_id.kind() != XTypes::TK_NONE) {
+      if (!writer_local || !reader_local) {
+        const DDS::DataRepresentationIdSeq repIds =
+          get_effective_data_rep_qos(tempDwQos.representation.value, false);
+        Encoding::Kind encoding_kind;
+        if (repr_to_encoding_kind(repIds[0], encoding_kind) && encoding_kind == Encoding::KIND_XCDR1) {
+          const XTypes::TypeFlag extensibility_mask = XTypes::IS_APPENDABLE;
+          if (type_lookup_service_->extensibility(extensibility_mask, writer_type_id)) {
+            if (DCPS_debug_level) {
+              ACE_DEBUG((LM_WARNING, ACE_TEXT("(%P|%t) WARNING: ")
+                ACE_TEXT("StaticEndpointManager::match_continue: ")
+                ACE_TEXT("Encountered unsupported combination of XCDR1 encoding and appendable extensibility\n")));
+            }
+          }
+        }
+      }
+
+      XTypes::TypeConsistencyAttributes type_consistency;
+      type_consistency.ignore_sequence_bounds = drQos->type_consistency.ignore_sequence_bounds;
+      type_consistency.ignore_string_bounds = drQos->type_consistency.ignore_string_bounds;
+      type_consistency.ignore_member_names = drQos->type_consistency.ignore_member_names;
+      type_consistency.prevent_type_widening = drQos->type_consistency.prevent_type_widening;
+      XTypes::TypeAssignability ta(type_lookup_service_, type_consistency);
+
+      if (drQos->type_consistency.kind == DDS::ALLOW_TYPE_COERCION) {
+        consistent = ta.assignable(reader_type_id, writer_type_id);
+      } else {
+        // The two types must be equivalent for DISALLOW_TYPE_COERCION
+        consistent = reader_type_id == writer_type_id;
+      }
+    } else {
+      if (drQos->type_consistency.force_type_validation) {
+        // Cannot do type validation since not both TypeObjects are available
+        consistent = false;
+      } else {
+        // Fall back to matching type names
+        OPENDDS_STRING writer_type_name;
+        OPENDDS_STRING reader_type_name;
+        if (writer_local) {
+          writer_type_name = td_iter->second.local_data_type_name();
+        } else {
+          writer_type_name = dpi->second.get_type_name();
+        }
+        if (reader_local) {
+          reader_type_name = td_iter->second.local_data_type_name();
+        } else {
+          reader_type_name = dsi->second.get_type_name();
+        }
+        consistent = writer_type_name == reader_type_name;
+      }
+    }
+
+    if (!consistent) {
+      td_iter->second.increment_inconsistent();
+      if (DCPS::DCPS_debug_level) {
+        ACE_DEBUG((LM_WARNING,
+                  ACE_TEXT("(%P|%t) StaticEndpointManager::match_continue - WARNING ")
+                  ACE_TEXT("Data types of topic %C does not match (inconsistent)\n"),
+                  topic_name.c_str()));
+      }
+      return;
+    }
+  }
+
+  // Need to release lock, below, for callbacks into DCPS which could
+  // call into Spdp/Sedp.  Note that this doesn't unlock, it just constructs
+  // an ACE object which will be used below for unlocking.
+  ACE_Reverse_Lock<ACE_Thread_Mutex> rev_lock(lock_);
+
+  // 4. Check transport and QoS compatibility
+
+  // Copy entries from local publication and local subscription maps
+  // prior to releasing lock
+  DataWriterCallbacks_wrch dwr;
+  DataReaderCallbacks_wrch drr;
+  if (writer_local) {
+    dwr = lpi->second.publication_;
+    OPENDDS_ASSERT(lpi->second.publication_);
+    OPENDDS_ASSERT(dwr);
+  }
+  if (reader_local) {
+    drr = lsi->second.subscription_;
+    OPENDDS_ASSERT(lsi->second.subscription_);
+    OPENDDS_ASSERT(drr);
+  }
+
+  IncompatibleQosStatus writerStatus = {0, 0, 0, DDS::QosPolicyCountSeq()};
+  IncompatibleQosStatus readerStatus = {0, 0, 0, DDS::QosPolicyCountSeq()};
+
+  if (compatibleQOS(&writerStatus, &readerStatus, *wTls, *rTls,
+      dwQos, drQos, pubQos, subQos)) {
+
+    bool call_writer = false, call_reader = false;
+
+    if (writer_local) {
+      call_writer = lpi->second.matched_endpoints_.insert(reader).second;
+      dwr = lpi->second.publication_;
+      if (!reader_local) {
+        dsi->second.matched_endpoints_.insert(writer);
+      }
+    }
+    if (reader_local) {
+      call_reader = lsi->second.matched_endpoints_.insert(writer).second;
+      drr = lsi->second.subscription_;
+      if (!writer_local) {
+        dpi->second.matched_endpoints_.insert(reader);
+      }
+    }
+
+    if (writer_local && !reader_local) {
+      add_assoc_i(writer, lpi->second, reader, dsi->second);
+    }
+    if (reader_local && !writer_local) {
+      add_assoc_i(reader, lsi->second, writer, dpi->second);
+    }
+
+    if (!call_writer && !call_reader) {
+      return; // nothing more to do
+    }
+
+    // Copy reader and writer association data prior to releasing lock
+    DDS::OctetSeq octet_seq_type_info_reader;
+    XTypes::serialize_type_info(*reader_type_info, octet_seq_type_info_reader);
+    const ReaderAssociation ra = {
+      *rTls, rTransportContext, reader, *subQos, *drQos,
+#ifndef OPENDDS_NO_CONTENT_FILTERED_TOPIC
+      cfProp->filterClassName, cfProp->filterExpression,
+#else
+      "", "",
+#endif
+      cfProp->expressionParameters,
+      octet_seq_type_info_reader,
+      reader_participant_discovered_at
+    };
+
+    DDS::OctetSeq octet_seq_type_info_writer;
+    XTypes::serialize_type_info(*writer_type_info, octet_seq_type_info_writer);
+    const WriterAssociation wa = {
+      *wTls, wTransportContext, writer, *pubQos, *dwQos,
+      octet_seq_type_info_writer,
+      writer_participant_discovered_at
+    };
+
+    ACE_GUARD(ACE_Reverse_Lock<ACE_Thread_Mutex>, rg, rev_lock);
+    static const bool writer_active = true;
+
+    if (call_writer) {
+      if (DCPS_debug_level > 3) {
+        ACE_DEBUG((LM_DEBUG, ACE_TEXT("(%P|%t) StaticEndpointManager::match_continue - ")
+          ACE_TEXT("adding writer %C association for reader %C\n"), OPENDDS_STRING(GuidConverter(writer)).c_str(),
+          OPENDDS_STRING(GuidConverter(reader)).c_str()));
+      }
+      DataWriterCallbacks_rch dwr_lock = dwr.lock();
+      if (dwr_lock) {
+        if (call_reader) {
+          DataReaderCallbacks_rch drr_lock = drr.lock();
+          if (drr_lock) {
+            DcpsUpcalls thr(drr_lock, reader, wa, !writer_active, dwr_lock);
+            thr.activate();
+            dwr_lock->add_association(writer, ra, writer_active);
+            thr.writer_done();
+          }
+        } else {
+          dwr_lock->add_association(writer, ra, writer_active);
+        }
+      }
+    } else if (call_reader) {
+      if (DCPS_debug_level > 3) {
+        ACE_DEBUG((LM_DEBUG, ACE_TEXT("(%P|%t) StaticEndpointManager::match_continue - ")
+          ACE_TEXT("adding reader %C association for writer %C\n"),
+          OPENDDS_STRING(GuidConverter(reader)).c_str(), OPENDDS_STRING(GuidConverter(writer)).c_str()));
+      }
+      DataReaderCallbacks_rch drr_lock = drr.lock();
+      if (drr_lock) {
+        drr_lock->add_association(reader, wa, !writer_active);
+      }
+    }
+
+  } else if (already_matched) { // break an existing associtaion
+    if (writer_local) {
+      lpi->second.matched_endpoints_.erase(reader);
+      lpi->second.remote_expectant_opendds_associations_.erase(reader);
+      if (dsi != discovered_subscriptions_.end()) {
+        dsi->second.matched_endpoints_.erase(writer);
+      }
+    }
+    if (reader_local) {
+      lsi->second.matched_endpoints_.erase(writer);
+      lsi->second.remote_expectant_opendds_associations_.erase(writer);
+      if (dpi != discovered_publications_.end()) {
+        dpi->second.matched_endpoints_.erase(reader);
+      }
+    }
+    if (writer_local && !reader_local) {
+      remove_assoc_i(writer, lpi->second, reader);
+    }
+    if (reader_local && !writer_local) {
+      remove_assoc_i(reader, lsi->second, writer);
+    }
+    ACE_GUARD(ACE_Reverse_Lock<ACE_Thread_Mutex>, rg, rev_lock);
+    if (writer_local) {
+      ReaderIdSeq reader_seq(1);
+      reader_seq.length(1);
+      reader_seq[0] = reader;
+      DataWriterCallbacks_rch dwr_lock = dwr.lock();
+      if (dwr_lock) {
+        dwr_lock->remove_associations(reader_seq, false /*notify_lost*/);
+      }
+    }
+    if (reader_local) {
+      WriterIdSeq writer_seq(1);
+      writer_seq.length(1);
+      writer_seq[0] = writer;
+      DataReaderCallbacks_rch drr_lock = drr.lock();
+      if (drr_lock) {
+        drr_lock->remove_associations(writer_seq, false /*notify_lost*/);
+      }
+    }
+  } else { // something was incompatible
+    ACE_GUARD(ACE_Reverse_Lock< ACE_Thread_Mutex>, rg, rev_lock);
+    if (writer_local && writerStatus.count_since_last_send) {
+      if (DCPS_debug_level > 3) {
+        ACE_DEBUG((LM_DEBUG, ACE_TEXT("(%P|%t) StaticEndpointManager::match - ")
+                   ACE_TEXT("writer incompatible\n")));
+      }
+      DataWriterCallbacks_rch dwr_lock = dwr.lock();
+      if (dwr_lock) {
+        dwr_lock->update_incompatible_qos(writerStatus);
+      }
+    }
+    if (reader_local && readerStatus.count_since_last_send) {
+      if (DCPS_debug_level > 3) {
+        ACE_DEBUG((LM_DEBUG, ACE_TEXT("(%P|%t) StaticEndpointManager::match - ")
+                   ACE_TEXT("reader incompatible\n")));
+      }
+      DataReaderCallbacks_rch drr_lock = drr.lock();
+      if (drr_lock) {
+        drr_lock->update_incompatible_qos(readerStatus);
+      }
+    }
+  }
+}
+
+void StaticEndpointManager::save_matching_data_and_get_typeobjects(
+  const XTypes::TypeInformation* type_info,
+  MatchingData& md, const MatchingPair& mp,
+  const GUID_t& remote_id,
+  bool is_discovery_protected)
+{
+  const SequenceNumber seqnum = ++type_lookup_service_sequence_number_;
+  if (DCPS_debug_level >= 4) {
+    ACE_DEBUG((LM_DEBUG,
+      "(%P|%t) StaticEndpointManager::save_matching_data_and_get_typeobjects: "
+      "remote: %C seq: %q\n", LogGuid(remote_id).c_str(), seqnum.getValue()));
+  }
+  md.rpc_sequence_number = seqnum;
+  MatchingDataIter md_it = matching_data_buffer_.find(mp);
+  if (md_it != matching_data_buffer_.end()) {
+    md_it->second = md;
+  } else {
+    matching_data_buffer_.insert(std::make_pair(mp, md));
+  }
+  // Store an entry for the first request
+  TypeIdOrigSeqNumber orig_req_data;
+  std::memcpy(orig_req_data.participant, remote_id.guidPrefix, sizeof(GuidPrefix_t));
+  orig_req_data.type_id = type_info->minimal.typeid_with_size.type_id;
+  orig_req_data.seq_number = md.rpc_sequence_number;
+  orig_req_data.secure = false;
+  orig_req_data.time_started = md.time_added_to_map;
+  orig_seq_numbers_.insert(std::make_pair(md.rpc_sequence_number, orig_req_data));
+
+  XTypes::TypeIdentifierSeq type_ids;
+  if (type_info->minimal.dependent_typeid_count == -1 ||
+      type_info->minimal.dependent_typeids.length() < (CORBA::ULong)type_info->minimal.dependent_typeid_count) {
+    type_ids.append(type_info->minimal.typeid_with_size.type_id);
+
+    // Get dependencies of topic type
+    send_type_lookup_request(type_ids, remote_id, is_discovery_protected, false);
+  } else {
+    type_ids.length(type_info->minimal.dependent_typeid_count + 1);
+    type_ids[0] = type_info->minimal.typeid_with_size.type_id;
+    for (unsigned i = 1; i <= (unsigned)type_info->minimal.dependent_typeid_count; ++i) {
+      type_ids[i] = type_info->minimal.dependent_typeids[i - 1].type_id;
+    }
+    // Get TypeObjects of topic type and all of its dependencies
+    send_type_lookup_request(type_ids, remote_id, is_discovery_protected, true);
+  }
+  type_lookup_reply_deadline_processor_->schedule(max_type_lookup_service_reply_period_);
+}
+
+GUID_t StaticEndpointManager::make_topic_guid()
+{
+  EntityId_t entity_id;
+  assign(entity_id.entityKey, topic_counter_);
+  ++topic_counter_;
+  entity_id.entityKind = ENTITYKIND_OPENDDS_TOPIC;
+
+  if (topic_counter_ == 0x1000000) {
+    ACE_ERROR((LM_ERROR,
+               ACE_TEXT("(%P|%t) ERROR: StaticEndpointManager::make_topic_guid: ")
+               ACE_TEXT("Exceeded Maximum number of topic entity keys!")
+               ACE_TEXT("Next key will be a duplicate!\n")));
+    topic_counter_ = 0;
+  }
+
+  return make_id(participant_id_, entity_id);
+}
+
+bool StaticEndpointManager::has_dcps_key(const GUID_t& topicId) const
+{
+  typedef OPENDDS_MAP_CMP(GUID_t, OPENDDS_STRING, GUID_tKeyLessThan) TNMap;
+  TNMap::const_iterator tn = topic_names_.find(topicId);
+  if (tn == topic_names_.end()) return false;
+
+  TopicDetailsMap::const_iterator td = topics_.find(tn->second);
+  if (td == topics_.end()) return false;
+
+  return td->second.has_dcps_key();
+}
+
 StaticDiscovery::StaticDiscovery(const RepoKey& key)
-  : PeerDiscovery<StaticParticipant>(key)
+  : Discovery(key)
 {}
 
 namespace {
@@ -1864,7 +2953,396 @@ void StaticDiscovery::pre_reader(DataReaderImpl* reader)
 
 StaticDiscovery_rch StaticDiscovery::instance_(make_rch<StaticDiscovery>(Discovery::DEFAULT_STATIC));
 
+DDS::Subscriber_ptr StaticDiscovery::init_bit(DomainParticipantImpl* participant)
+{
+  DDS::Subscriber_var bit_subscriber;
+#ifndef DDS_HAS_MINIMUM_BIT
+  if (!TheServiceParticipant->get_BIT()) {
+    get_part(participant->get_domain_id(), participant->get_id())->init_bit(bit_subscriber);
+    return 0;
+  }
+
+  if (create_bit_topics(participant) != DDS::RETCODE_OK) {
+    return 0;
+  }
+
+  bit_subscriber =
+    participant->create_subscriber(SUBSCRIBER_QOS_DEFAULT,
+                                   DDS::SubscriberListener::_nil(),
+                                   DEFAULT_STATUS_MASK);
+  SubscriberImpl* sub = dynamic_cast<SubscriberImpl*>(bit_subscriber.in());
+  if (sub == 0) {
+    ACE_ERROR((LM_ERROR, ACE_TEXT("(%P|%t) PeerDiscovery::init_bit")
+               ACE_TEXT(" - Could not cast Subscriber to SubscriberImpl\n")));
+    return 0;
+  }
+
+  DDS::DataReaderQos dr_qos;
+  sub->get_default_datareader_qos(dr_qos);
+  dr_qos.durability.kind = DDS::TRANSIENT_LOCAL_DURABILITY_QOS;
+
+  dr_qos.reader_data_lifecycle.autopurge_nowriter_samples_delay =
+    TheServiceParticipant->bit_autopurge_nowriter_samples_delay();
+  dr_qos.reader_data_lifecycle.autopurge_disposed_samples_delay =
+    TheServiceParticipant->bit_autopurge_disposed_samples_delay();
+
+  DDS::TopicDescription_var bit_part_topic =
+    participant->lookup_topicdescription(BUILT_IN_PARTICIPANT_TOPIC);
+  create_bit_dr(bit_part_topic, BUILT_IN_PARTICIPANT_TOPIC_TYPE,
+                sub, dr_qos);
+
+  DDS::TopicDescription_var bit_topic_topic =
+    participant->lookup_topicdescription(BUILT_IN_TOPIC_TOPIC);
+  create_bit_dr(bit_topic_topic, BUILT_IN_TOPIC_TOPIC_TYPE,
+                sub, dr_qos);
+
+  DDS::TopicDescription_var bit_pub_topic =
+    participant->lookup_topicdescription(BUILT_IN_PUBLICATION_TOPIC);
+  create_bit_dr(bit_pub_topic, BUILT_IN_PUBLICATION_TOPIC_TYPE,
+                sub, dr_qos);
+
+  DDS::TopicDescription_var bit_sub_topic =
+    participant->lookup_topicdescription(BUILT_IN_SUBSCRIPTION_TOPIC);
+  create_bit_dr(bit_sub_topic, BUILT_IN_SUBSCRIPTION_TOPIC_TYPE,
+                sub, dr_qos);
+
+  DDS::TopicDescription_var bit_part_loc_topic =
+    participant->lookup_topicdescription(BUILT_IN_PARTICIPANT_LOCATION_TOPIC);
+  create_bit_dr(bit_part_loc_topic, BUILT_IN_PARTICIPANT_LOCATION_TOPIC_TYPE,
+                sub, dr_qos);
+
+  DDS::TopicDescription_var bit_connection_record_topic =
+    participant->lookup_topicdescription(BUILT_IN_CONNECTION_RECORD_TOPIC);
+  create_bit_dr(bit_connection_record_topic, BUILT_IN_CONNECTION_RECORD_TOPIC_TYPE,
+                sub, dr_qos);
+
+  DDS::TopicDescription_var bit_internal_thread_topic =
+    participant->lookup_topicdescription(BUILT_IN_INTERNAL_THREAD_TOPIC);
+  create_bit_dr(bit_internal_thread_topic, BUILT_IN_INTERNAL_THREAD_TOPIC_TYPE,
+                sub, dr_qos);
+
+  const DDS::ReturnCode_t ret = bit_subscriber->enable();
+  if (ret != DDS::RETCODE_OK) {
+    if (DCPS_debug_level) {
+      ACE_DEBUG((LM_INFO, ACE_TEXT("(%P|%t) PeerDiscovery::init_bit")
+                 ACE_TEXT(" - Error %d enabling subscriber\n"), ret));
+    }
+    return 0;
+  }
+#endif /* DDS_HAS_MINIMUM_BIT */
+
+  get_part(participant->get_domain_id(), participant->get_id())->init_bit(bit_subscriber);
+
+  return bit_subscriber._retn();
 }
+
+void StaticDiscovery::fini_bit(DCPS::DomainParticipantImpl* participant)
+{
+  get_part(participant->get_domain_id(), participant->get_id())->fini_bit();
 }
+
+bool StaticDiscovery::attach_participant(
+  DDS::DomainId_t /*domainId*/, const GUID_t& /*participantId*/)
+{
+  return false; // This is just for DCPSInfoRepo?
+}
+
+bool StaticDiscovery::remove_domain_participant(
+  DDS::DomainId_t domain_id, const GUID_t& participantId)
+{
+  // Use reference counting to ensure participant
+  // does not get deleted until lock as been released.
+  ParticipantHandle participant;
+  ACE_GUARD_RETURN(ACE_Thread_Mutex, g, lock_, false);
+  typename DomainParticipantMap::iterator domain = participants_.find(domain_id);
+  if (domain == participants_.end()) {
+    return false;
+  }
+  typename ParticipantMap::iterator part = domain->second.find(participantId);
+  if (part == domain->second.end()) {
+    return false;
+  }
+  participant = part->second;
+  domain->second.erase(part);
+  if (domain->second.empty()) {
+    participants_.erase(domain);
+  }
+
+  participant->shutdown();
+  return true;
+}
+
+bool StaticDiscovery::ignore_domain_participant(
+  DDS::DomainId_t domain, const GUID_t& myParticipantId, const GUID_t& ignoreId)
+{
+  get_part(domain, myParticipantId)->ignore_domain_participant(ignoreId);
+  return true;
+}
+
+bool StaticDiscovery::update_domain_participant_qos(
+  DDS::DomainId_t domain, const GUID_t& participant, const DDS::DomainParticipantQos& qos)
+{
+  return get_part(domain, participant)->update_domain_participant_qos(qos);
+}
+
+DCPS::TopicStatus StaticDiscovery::assert_topic(
+  GUID_t& topicId,
+  DDS::DomainId_t domainId,
+  const GUID_t& participantId,
+  const char* topicName,
+  const char* dataTypeName,
+  const DDS::TopicQos& qos,
+  bool hasDcpsKey,
+  DCPS::TopicCallbacks* topic_callbacks)
+{
+  ACE_GUARD_RETURN(ACE_Thread_Mutex, g, lock_, DCPS::INTERNAL_ERROR);
+  // Verified its safe to hold lock during call to assert_topic
+  return participants_[domainId][participantId]->assert_topic(topicId, topicName,
+                                                              dataTypeName, qos,
+                                                              hasDcpsKey, topic_callbacks);
+}
+
+DCPS::TopicStatus StaticDiscovery::find_topic(
+  DDS::DomainId_t domainId,
+  const GUID_t& participantId,
+  const char* topicName,
+  CORBA::String_out dataTypeName,
+  DDS::TopicQos_out qos,
+  GUID_t& topicId)
+{
+  ACE_GUARD_RETURN(ACE_Thread_Mutex, g, lock_, DCPS::INTERNAL_ERROR);
+  return participants_[domainId][participantId]->find_topic(topicName, dataTypeName, qos, topicId);
+}
+
+DCPS::TopicStatus StaticDiscovery::remove_topic(
+  DDS::DomainId_t domainId,
+  const GUID_t& participantId,
+  const GUID_t& topicId)
+{
+  ACE_GUARD_RETURN(ACE_Thread_Mutex, g, lock_, DCPS::INTERNAL_ERROR);
+  // Safe to hold lock while calling remove topic
+  return participants_[domainId][participantId]->remove_topic(topicId);
+}
+
+bool StaticDiscovery::ignore_topic(DDS::DomainId_t domainId, const GUID_t& myParticipantId,
+                                 const GUID_t& ignoreId)
+{
+  get_part(domainId, myParticipantId)->ignore_topic(ignoreId);
+  return true;
+}
+
+bool StaticDiscovery::update_topic_qos(const GUID_t& topicId, DDS::DomainId_t domainId,
+                                    const GUID_t& participantId, const DDS::TopicQos& qos)
+{
+  ACE_GUARD_RETURN(ACE_Thread_Mutex, g, lock_, false);
+  // Safe to hold lock while calling update_topic_qos
+  return participants_[domainId][participantId]->update_topic_qos(topicId, qos);
+}
+
+GUID_t StaticDiscovery::add_publication(
+  DDS::DomainId_t domainId,
+  const GUID_t& participantId,
+  const GUID_t& topicId,
+  DCPS::DataWriterCallbacks_rch publication,
+  const DDS::DataWriterQos& qos,
+  const DCPS::TransportLocatorSeq& transInfo,
+  const DDS::PublisherQos& publisherQos,
+  const XTypes::TypeInformation& type_info)
+{
+  return get_part(domainId, participantId)->add_publication(
+    topicId, publication, qos, transInfo, publisherQos, type_info);
+}
+
+bool StaticDiscovery::remove_publication(
+  DDS::DomainId_t domainId, const GUID_t& participantId, const GUID_t& publicationId)
+{
+  get_part(domainId, participantId)->remove_publication(publicationId);
+  return true;
+}
+
+bool StaticDiscovery::ignore_publication(
+  DDS::DomainId_t domainId, const GUID_t& participantId, const GUID_t& ignoreId)
+{
+  get_part(domainId, participantId)->ignore_publication(ignoreId);
+  return true;
+}
+
+bool StaticDiscovery::update_publication_qos(
+  DDS::DomainId_t domainId,
+  const GUID_t& partId,
+  const GUID_t& dwId,
+  const DDS::DataWriterQos& qos,
+  const DDS::PublisherQos& publisherQos)
+{
+  return get_part(domainId, partId)->update_publication_qos(dwId, qos,
+                                                            publisherQos);
+}
+
+void StaticDiscovery::update_publication_locators(
+  DDS::DomainId_t domainId, const GUID_t& partId, const GUID_t& dwId,
+  const DCPS::TransportLocatorSeq& transInfo)
+{
+  get_part(domainId, partId)->update_publication_locators(dwId, transInfo);
+}
+
+GUID_t StaticDiscovery::add_subscription(
+  DDS::DomainId_t domainId,
+  const GUID_t& participantId,
+  const GUID_t& topicId,
+  DCPS::DataReaderCallbacks_rch subscription,
+  const DDS::DataReaderQos& qos,
+  const DCPS::TransportLocatorSeq& transInfo,
+  const DDS::SubscriberQos& subscriberQos,
+  const char* filterClassName,
+  const char* filterExpr,
+  const DDS::StringSeq& params,
+  const XTypes::TypeInformation& type_info)
+{
+  return get_part(domainId, participantId)->add_subscription(
+    topicId, subscription, qos, transInfo, subscriberQos, filterClassName,
+    filterExpr, params, type_info);
+}
+
+bool StaticDiscovery::remove_subscription(
+  DDS::DomainId_t domainId, const GUID_t& participantId, const GUID_t& subscriptionId)
+{
+  get_part(domainId, participantId)->remove_subscription(subscriptionId);
+  return true;
+}
+
+bool StaticDiscovery::ignore_subscription(
+  DDS::DomainId_t domainId, const GUID_t& participantId, const GUID_t& ignoreId)
+{
+  get_part(domainId, participantId)->ignore_subscription(ignoreId);
+  return true;
+}
+
+bool StaticDiscovery::update_subscription_qos(
+  DDS::DomainId_t domainId,
+  const GUID_t& partId,
+  const GUID_t& drId,
+  const DDS::DataReaderQos& qos,
+  const DDS::SubscriberQos& subQos)
+{
+  return get_part(domainId, partId)->update_subscription_qos(drId, qos, subQos);
+}
+
+bool StaticDiscovery::update_subscription_params(
+  DDS::DomainId_t domainId, const GUID_t& partId, const GUID_t& subId, const DDS::StringSeq& params)
+{
+  return get_part(domainId, partId)->update_subscription_params(subId, params);
+}
+
+void StaticDiscovery::update_subscription_locators(
+  DDS::DomainId_t domainId, const GUID_t& partId, const GUID_t& subId,
+  const DCPS::TransportLocatorSeq& transInfo)
+{
+  get_part(domainId, partId)->update_subscription_locators(subId, transInfo);
+}
+
+StaticDiscovery::ParticipantHandle StaticDiscovery::get_part(
+  const DDS::DomainId_t domain_id, const GUID_t& part_id) const
+{
+  ACE_GUARD_RETURN(ACE_Thread_Mutex, g, lock_, ParticipantHandle());
+  typename DomainParticipantMap::const_iterator domain = participants_.find(domain_id);
+  if (domain == participants_.end()) {
+    return ParticipantHandle();
+  }
+  typename ParticipantMap::const_iterator part = domain->second.find(part_id);
+  if (part == domain->second.end()) {
+    return ParticipantHandle();
+  }
+  return part->second;
+}
+
+void StaticDiscovery::create_bit_dr(DDS::TopicDescription_ptr topic, const char* type,
+  SubscriberImpl* sub, const DDS::DataReaderQos& qos)
+{
+  TopicDescriptionImpl* bit_topic_i =
+    dynamic_cast<TopicDescriptionImpl*>(topic);
+  if (bit_topic_i == 0) {
+    ACE_ERROR((LM_ERROR, ACE_TEXT("(%P|%t) PeerDiscovery::create_bit_dr")
+               ACE_TEXT(" - Could not cast TopicDescription to TopicDescriptionImpl\n")));
+    return;
+  }
+
+  DDS::DomainParticipant_var participant = sub->get_participant();
+  DomainParticipantImpl* participant_i =
+    dynamic_cast<DomainParticipantImpl*>(participant.in());
+  if (participant_i == 0) {
+    ACE_ERROR((LM_ERROR, ACE_TEXT("(%P|%t) PeerDiscovery::create_bit_dr")
+               ACE_TEXT(" - Could not cast DomainParticipant to DomainParticipantImpl\n")));
+    return;
+  }
+
+  TypeSupport_var type_support =
+    Registered_Data_Types->lookup(participant, type);
+
+  DDS::DataReader_var dr = type_support->create_datareader();
+  DataReaderImpl* dri = dynamic_cast<DataReaderImpl*>(dr.in());
+  if (dri == 0) {
+    ACE_ERROR((LM_ERROR, ACE_TEXT("(%P|%t) PeerDiscovery::create_bit_dr")
+               ACE_TEXT(" - Could not cast DataReader to DataReaderImpl\n")));
+    return;
+  }
+
+  dri->init(bit_topic_i, qos, 0 /*listener*/, 0 /*mask*/, participant_i, sub);
+  dri->disable_transport();
+  dri->enable();
+}
+
+void StaticParticipant::remove_discovered_participant(DiscoveredParticipantIter& iter)
+{
+  if (iter == participants_.end()) {
+    return;
+  }
+  GUID_t part_id = iter->first;
+  bool removed = endpoint_manager().disassociate();
+  iter = participants_.find(part_id); // refresh iter after disassociate, which can unlock
+  if (iter == participants_.end()) {
+    return;
+  }
+  if (removed) {
+#ifndef DDS_HAS_MINIMUM_BIT
+    ParticipantBuiltinTopicDataDataReaderImpl* bit = part_bit();
+    ParticipantLocationBuiltinTopicDataDataReaderImpl* loc_bit = part_loc_bit();
+    // bit may be null if the DomainParticipant is shutting down
+    if ((bit && iter->second.bit_ih_ != DDS::HANDLE_NIL) ||
+        (loc_bit && iter->second.location_ih_ != DDS::HANDLE_NIL)) {
+      {
+        const DDS::InstanceHandle_t bit_ih = iter->second.bit_ih_;
+        const DDS::InstanceHandle_t location_ih = iter->second.location_ih_;
+
+        ACE_Reverse_Lock<ACE_Thread_Mutex> rev_lock(lock_);
+        ACE_GUARD(ACE_Reverse_Lock<ACE_Thread_Mutex>, rg, rev_lock);
+        if (bit && bit_ih != DDS::HANDLE_NIL) {
+          bit->set_instance_state(bit_ih,
+                                  DDS::NOT_ALIVE_DISPOSED_INSTANCE_STATE);
+        }
+        if (loc_bit && location_ih != DDS::HANDLE_NIL) {
+          loc_bit->set_instance_state(location_ih,
+                                      DDS::NOT_ALIVE_DISPOSED_INSTANCE_STATE);
+        }
+      }
+      iter = participants_.find(part_id);
+      if (iter == participants_.end()) {
+        return;
+      }
+    }
+#endif /* DDS_HAS_MINIMUM_BIT */
+    if (DCPS_debug_level > 3) {
+      GuidConverter conv(iter->first);
+      ACE_DEBUG((LM_INFO, ACE_TEXT("(%P|%t) LocalParticipant::remove_discovered_participant")
+                 ACE_TEXT(" - erasing %C (%B)\n"), OPENDDS_STRING(conv).c_str(), participants_.size()));
+    }
+
+    remove_discovered_participant_i(iter);
+
+    participants_.erase(iter);
+  }
+}
+
+} // namespace DCPS
+} // namespace OpenDDS
 
 OPENDDS_END_VERSIONED_NAMESPACE_DECL
