+++ tests/DCPS/BitDataReader/BitDataReader.cpp
@@ -1,22 +1,22 @@
-
-#include "ace/OS_main.h"
 #include "../common/TestSupport.h"
-#include "dds/DdsDcpsDomainC.h"
-#include "dds/DdsDcpsCoreTypeSupportImpl.h"
-#include "dds/DCPS/Service_Participant.h"
-#include "dds/DCPS/Marked_Default_Qos.h"
-#include "dds/DCPS/BuiltInTopicUtils.h"
-#include "dds/DCPS/WaitSet.h"
-#include "dds/DCPS/DiscoveryBase.h"
 
-#include "dds/DCPS/StaticIncludes.h"
+#include <dds/DCPS/Service_Participant.h>
+#include <dds/DCPS/Marked_Default_Qos.h>
+#include <dds/DCPS/BuiltInTopicUtils.h>
+#include <dds/DCPS/BuiltInTopicDataReaderImpls.h>
+#include <dds/DCPS/WaitSet.h>
+#include <dds/DCPS/StaticIncludes.h>
 #ifdef ACE_AS_STATIC_LIBS
-#include "dds/DCPS/RTPS/RtpsDiscovery.h"
-#include "dds/DCPS/transport/rtps_udp/RtpsUdp.h"
+#  include <dds/DCPS/RTPS/RtpsDiscovery.h>
+#  include <dds/DCPS/transport/rtps_udp/RtpsUdp.h>
 #endif
 
-#include "ace/OS_NS_unistd.h"
-#include "ace/Log_Msg.h"
+#include <dds/DdsDcpsDomainC.h>
+#include <dds/DdsDcpsCoreTypeSupportImpl.h>
+
+#include <ace/OS_main.h>
+#include <ace/OS_NS_unistd.h>
+#include <ace/Log_Msg.h>
 
 using namespace DDS;
 using namespace OpenDDS::DCPS;
