+++ tests/DCPS/common/ConnectionRecordLogger.cpp
@@ -1,9 +1,8 @@
 #include "ConnectionRecordLogger.h"
 
-#include "dds/DCPS/BuiltInTopicUtils.h"
-#include "dds/DCPS/DiscoveryBase.h"
-#include "dds/DCPS/GuidConverter.h"
-#include "dds/DCPS/JsonValueWriter.h"
+#include <dds/DCPS/BuiltInTopicUtils.h>
+#include <dds/DCPS/GuidConverter.h>
+#include <dds/DCPS/JsonValueWriter.h>
 
 OPENDDS_BEGIN_VERSIONED_NAMESPACE_DECL
